Update entity-prop-types.md - Round #1
The biggest thing here is just double checking the words in fancy ticks to make sure they have a "?" or "$" when needed, or that everything is capitalized (or not) appropriately. There were A LOT of them, so there's a good chance I missed one or messed it up in a place or ten. Changed some wording around here and there too, so you'll just need to make sure it still makes sense.